requirements-alt.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2==2.9.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
python-telegram-bot==20.7
twilio==8.10.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
